feat(cli): share spec-discovery step across fern mcp init and tools
