+++ packages/cli/cli/src/commands/mcp/prototype/initMcp.ts
@@ -3,7 +3,6 @@ import { AbsoluteFilePath } from "@fern-api/fs-utils";
 import { Project } from "@fern-api/project-loader";
 import { select } from "@inquirer/prompts";
 import chalk from "chalk";
-import path from "path";
 
 import { CliContext } from "../../../cli-context/CliContext.js";
 import { proposeAiRuleset } from "./aiCurated.js";
@@ -18,10 +17,9 @@ import { BuiltinPresetKey, buildBuiltinPresets, PresetResolution } from "./prese
 import { computeVerdict, formatTokens, formatVerdictLine, resolveTools, ToolsConfig } from "./toolset.js";
 import { runTrimLoop } from "./trimLoop.js";
 import { banner, command, hint, ICONS, radioChoice, selectTheme, sleep, styledVerdictLine, withSpinner } from "./ui.js";
-import { pickWorkspaceAndLoadSpec } from "./workspace.js";
+import { announceSpecDiscovery, pickWorkspaceAndLoadSpec } from "./workspace.js";
 
 const AI_SPINNER_DELAY_MS = 1200;
-const SPEC_SCAN_DELAY_MS = 600;
 
 export interface InitMcpArgs {
     project: Project;
@@ -105,11 +103,6 @@ export async function initMcp(args: InitMcpArgs): Promise<void> {
     const { project, cliContext, yes, json } = args;
     const interactive = !yes && !json;
 
-    if (!json) {
-        cliContext.logger.info("");
-        await withSpinner("Scanning workspace for API specs…", () => sleep(SPEC_SCAN_DELAY_MS));
-    }
-
     const workspaceSpec = await pickWorkspaceAndLoadSpec({
         project,
         cliContext,
@@ -119,6 +112,9 @@ export async function initMcp(args: InitMcpArgs): Promise<void> {
     if (workspaceSpec == null) {
         return;
     }
+    if (!json) {
+        await announceSpecDiscovery({ cliContext, workspaceSpec });
+    }
     const { spec, workspaceName, absolutePathToWorkspace } = workspaceSpec;
     const endpoints = spec.endpoints;
     const allTools = resolveTools(endpoints, {});
@@ -127,10 +123,6 @@ export async function initMcp(args: InitMcpArgs): Promise<void> {
     if (!json) {
         const title = spec.title ?? workspaceName;
         const tagless = endpoints.every((endpoint) => endpoint.tags.length === 0);
-        const relativeSpecPath = path.relative(absolutePathToWorkspace, spec.absoluteFilePath);
-        cliContext.logger.info(
-            `${ICONS.success} Found spec ${chalk.bold(relativeSpecPath)} ${hint(`${ICONS.bullet} ${title} ${ICONS.bullet} ${endpoints.length} endpoints`)}`
-        );
         cliContext.logger.info("");
         cliContext.logger.info(
             banner("Create an MCP server", [
